fix(streamwall): key overlay view tiles by their grid position (#39)
Each active view's border/tile in the transparent overlay renderer
was rendered without a `key`, so Preact matched them purely by array
position. When an earlier view disappeared, every later view's
position shifted down a slot and Preact grafted the surviving view's
content onto a DOM node that belonged to a different view.

Extracted Overlay (and its supporting VersionFooter/styled-components)
out of overlay.tsx into OverlayRoot.tsx, mirroring the existing
OverlayViewTile.tsx split, so it can be rendered and tested without
pulling in the module-level `render(<App />, document.body)` bootstrap.
Keyed each tile by its view's stable grid position (`pos.spaces[0]`).

Added OverlayRoot.test.tsx, which renders Overlay directly and asserts
a surviving tile's DOM node identity is preserved when an earlier view
disappears; it fails without the key. Updated
layerFrameSandbox.sources.test.ts's file list for the move.
